refactor: 커스텀 훅 수정
